Moved the StableDiffusion generation to this repo.
Had to make some tweaks.
Removed a bunch from the old illustrator_agent and placed the slimmed
down version in stable_diffusion.py
We are now only taking in a prompt to the class and spitting out an
image.
Currently no API implementation: run test.py to generate images locally.