Removed model relocation to CPU on cleanup
The model was moved to CPU before deletion.
This is common practice but our model doesnt even support cpu usage so
it gave us a warning on cleanup.
I deleted the line where we moved it to CPU before deletion.
Shouldnt make a difference since we are deleting everything in the
python garbage collector anyways.